+++ sites/public/pages/applications/household/programs.tsx
@@ -10,7 +10,6 @@ import {
   t,
   Button,
   AppearanceStyleType,
-  AuthContext,
 } from "@bloom-housing/ui-components"
 import FormsLayout from "../../../layouts/forms"
 import FormBackLink from "../../../src/forms/applications/FormBackLink"
@@ -22,6 +21,7 @@ import {
   OnClientSide,
   PageView,
   pushGtmEvent,
+  AuthContext,
 } from "@bloom-housing/shared-helpers"
 import { FormMetaDataType } from "@bloom-housing/backend-core/types"
 import { UserStatus } from "../../../lib/constants"