+++ ui-components/index.ts
@@ -19,7 +19,6 @@ export * from "./src/blocks/ViewItem"
 export * from "./src/blocks/StandardCard"
 
 /* Contexts */
-export * from "./src/authentication"
 export * from "./src/config"
 
 /* Footers */
@@ -65,7 +64,6 @@ export * from "./src/helpers/numberOrdinal"
 export * from "./src/helpers/OneLineAddress"
 export * from "./src/helpers/preferences"
 export * from "./src/helpers/resolveObject"
-export * from "./src/helpers/tableSummaries"
 export * from "./src/helpers/translator"
 export * from "./src/helpers/useIntersect"
 export * from "./src/helpers/useMutate"
@@ -87,11 +85,14 @@ export * from "./src/navigation/Tabs"
 export * from "./src/navigation/Breadcrumbs"
 
 /* Notifications */
+export * from "./src/notifications/AlertBox"
+export * from "./src/notifications/AlertNotice"
 export * from "./src/notifications/ApplicationStatus"
 export * from "./src/notifications/ErrorMessage"
+export * from "./src/notifications/SiteAlert"
 export * from "./src/notifications/StatusAside"
 export * from "./src/notifications/StatusMessage"
-export * from "./src/notifications"
+export * from "./src/notifications/alertTypes"
 
 /* Overlays */
 export * from "./src/overlays/Overlay"
@@ -108,7 +109,6 @@ export * from "./src/page_components/listing/ListingDetailHeader"
 export * from "./src/page_components/listing/ListingDetails"
 export * from "./src/page_components/listing/ListingMap"
 export * from "./src/page_components/listing/ListingsGroup"
-export * from "./src/page_components/listing/UnitTables"
 export * from "./src/page_components/listing/listing_sidebar/GetApplication"
 export * from "./src/page_components/listing/listing_sidebar/Contact"
 export * from "./src/page_components/listing/listing_sidebar/ListingUpdated"
@@ -118,7 +118,7 @@ export * from "./src/page_components/listing/listing_sidebar/ReferralApplication
 export * from "./src/page_components/listing/listing_sidebar/ContactAddress"
 export * from "./src/page_components/listing/listing_sidebar/SubmitApplication"
 export * from "./src/page_components/listing/listing_sidebar/QuantityRowSection"
-export * from "./src/page_components/listing/listing_sidebar/WhatToExpect"
+export * from "./src/page_components/listing/listing_sidebar/ExpandableSection"
 export * from "./src/page_components/listing/listing_sidebar/events/DownloadLotteryResults"
 export * from "./src/page_components/listing/listing_sidebar/events/EventSection"
 export * from "./src/page_components/sign-in/FormTerms"